Put all together for better source code in storybook
