--- stories/Templates/homepage/Footer.js
@@ -1,241 +1,0 @@
-import React from 'react';
-import {Container, Row, Col, Icon} from '../../../src'
-
-const Footer = () => <footer className="it-footer">
-    <div className="it-footer-main">
-        <Container>
-            <section>
-                <Row className="clearfix">
-                    <Col sm={12}>
-                        <div className="it-brand-wrapper">
-                            <a href="#" className="">
-                                <Icon icon="it-pa" />
-                                <div className="it-brand-text">
-                                    <h2 className="no_toc">Nome del Comune</h2>
-                                    <h3 className="no_toc d-none d-md-block">
-                                        Uno dei tanti Comuni d'Italia
-      </h3>
-                                </div>
-                            </a>
-                        </div>
-                    </Col>
-                </Row>
-            </section>
-            <section>
-                <Row>
-                    <Col lg={3} md={3} sm={6} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Amministrazione">Amministrazione</a>
-                        </h4>
-                        <div className="link-list-wrapper">
-                            <ul className="footer-list link-list clearfix">
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Giunta e consiglio">
-                                        Giunta e consiglio
-                                </a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Aree di competenza">
-                                        Aree di competenza
-                                </a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Dipendenti">Dipendenti</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Luoghi">Luoghi</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Associazioni e società partecipate">
-                                        Associazioni e società partecipate
-                                </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </Col>
-                    <Col lg={3} md={3} sm={6} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Servizi">Servizi</a>
-                        </h4>
-                        <div className="link-list-wrapper">
-                            <ul className="footer-list link-list clearfix">
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Pagamenti">Pagamenti</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Sostegno">Sostegno</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Domande e iscrizioni">
-                                        Domande e iscrizioni
-                                </a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Segnalazioni">Segnalazioni</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Autorizzazioni e concessioni">
-                                        Autorizzazioni e concessioni
-                                </a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Certificati e dichiarazioni">
-                                        Certificati e dichiarazioni
-                                </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </Col>
-                    <Col lg={3} md={3} sm={6} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Novità">Novità</a>
-                        </h4>
-                        <div className="link-list-wrapper">
-                            <ul className="footer-list link-list clearfix">
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Notizie">Notizie</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Eventi">Eventi</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Comunicati stampa">
-                                        Comunicati stampa
-                                </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </Col>
-                    <Col lg={3} md={3} sm={6}>
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Documenti">Documenti</a>
-                        </h4>
-                        <div className="link-list-wrapper">
-                            <ul className="footer-list link-list clearfix">
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Progetti e attività">Progetti e
-        attività</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Delibere, determine e ordinanze">Delibere,
-        determine e ordinanze</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Bandi">Bandi</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Concorsi">Concorsi</a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Albo pretorio">Albo pretorio</a>
-                                </li>
-                            </ul>
-                        </div>
-                    </Col>
-                </Row>
-            </section>
-            <section className="py-4 border-white border-top">
-                <Row>
-                    <Col lg={3} md={3} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Amministrazione">
-                                Amministrazione trasparente
-                        </a>
-                        </h4>
-                        <p>
-                            I dati personali pubblicati sono riutilizzabili solo alle
-                            condizioni previste dalla direttiva comunitaria 2003/98/CE e dal
-                            d.lgs. 36/2006
-                    </p>
-                    </Col>
-                    <Col lg={3} md={3} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Contatti">Contatti</a>
-                        </h4>
-                        <p>
-                            <strong>Nome del Comune</strong><br />
-                            Via Roma 0 - 00000 Lorem Ipsum Codice fiscale / P. IVA: 000000000
-                    </p>
-                        <div className="link-list-wrapper">
-                            <ul className="footer-list link-list clearfix">
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: Posta Elettronica Certificata">
-                                        Posta Elettronica Certificata
-                                </a>
-                                </li>
-                                <li>
-                                    <a className="list-item" href="#" title="Vai alla pagina: URP - Ufficio Relazioni con il Pubblico">
-                                        URP - Ufficio Relazioni con il Pubblico
-                                </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </Col>
-                    <Col lg={3} md={3} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Newsletter">Newsletter</a>
-                        </h4>
-                        <form action="#" className="form-newsletter" method="post">
-                            <label className="text-white font-weight-semibold active" htmlFor="input-newsletter" style={{ transition: "none 0s ease 0s" }}>Iscriviti per riceverla</label>
-                            <input className="form-control" id="input-newsletter" name="input-newsletter" placeholder="mail@example.com" type="email" />
-                            <button className="btn btn-primary btn-icon" type="submit">
-                                <Icon icon='it-mail' color="white" />
-                                <span>Iscriviti</span>
-                            </button>
-                        </form>
-                    </Col>
-                    <Col lg={3} md={3} className="pb-2">
-                        <h4>
-                            <a href="#" title="Vai alla pagina: Seguici su">Seguici su</a>
-                        </h4>
-                        <ul className="list-inline text-left social">
-                            <li className="list-inline-item">
-                                <a className="p-2 text-white" href="#" target="_blank">
-                                    <Icon icon='it-designers-italia' color="white" size="sm" className="align-top" />
-                                    <span className="sr-only">Designers Italia</span></a>
-                            </li>
-                            <li className="list-inline-item">
-                                <a className="p-2 text-white" href="#" target="_blank">
-
-                                    <Icon icon='it-twitter' color="white" size="sm" className="align-top" />
-                                    <span className="sr-only">Twitter</span></a>
-                            </li>
-                            <li className="list-inline-item">
-                                <a className="p-2 text-white" href="#" target="_blank">
-
-                                    <Icon icon='it-medium' color="white" size="sm" className="align-top" />
-                                    <span className="sr-only">Medium</span></a>
-                            </li>
-                            <li className="list-inline-item">
-                                <a className="p-2 text-white" href="#" target="_blank">
-                                    <Icon icon='it-behance' color="white" size="sm" className="align-top" />
-                                    <span className="sr-only">Behance</span></a>
-                            </li>
-                        </ul>
-                    </Col>
-                </Row>
-            </section>
-        </Container>
-    </div>
-    <div className="it-footer-small-prints clearfix">
-        <Container>
-            <h3 className="sr-only">Sezione Link Utili</h3>
-            <ul className="it-footer-small-prints-list list-inline mb-0 d-flex flex-column flex-md-row">
-                <li className="list-inline-item">
-                    <a href="#" title="Note Legali">Media policy</a>
-                </li>
-                <li className="list-inline-item">
-                    <a href="#" title="Note Legali">Note legali</a>
-                </li>
-                <li className="list-inline-item">
-                    <a href="#" title="Privacy-Cookies">Privacy policy</a>
-                </li>
-                <li className="list-inline-item">
-                    <a href="#" title="Mappa del sito">Mappa del sito</a>
-                </li>
-            </ul>
-        </Container>
-    </div>
-</footer>
-
-export default Footer;
